add round duration setting and cleanup

+++ Assets/Scripts/Data/GameSettingsData.cs
@@ -7,5 +7,5 @@ using UnityEngine;
 public class GameSettingsData
 {
     public string wordsFile;
-    public float roundDuration;
+    public int roundDuration;
 }
